Extract gcode functions from dialog
