+++ src/renderer/src/features/gcode-options/gcodePrefsValidation.ts
@@ -1,0 +1,11 @@
+export function parsePositiveNumber(value: string): number | null {
+  const parsed = Number.parseFloat(value);
+  if (!Number.isFinite(parsed) || parsed <= 0) return null;
+  return parsed;
+}
+
+export function parseNonNegativeNumber(value: string): number | null {
+  const parsed = Number.parseFloat(value);
+  if (!Number.isFinite(parsed) || parsed < 0) return null;
+  return parsed;
+}
